+++ internal/Panels/PopupDecorator.js
@@ -63,7 +63,7 @@ const defaultConfig = {
  */
 const PopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 	const {className: cfgClassName, css, panelArranger, panelType} = config;
-	const Panels = CancelDecorator({cancel: 'onBack'}, Wrapped);
+	const CancelableWrapped = CancelDecorator({cancel: 'onBack'}, Wrapped);
 
 	const Decorator = kind({
 		name: 'PopupDecorator',
@@ -83,6 +83,18 @@ const PopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			 * @type {Node}
 			 */
 			children: PropTypes.node,
+
+			/**
+			 * Instructs the Popup to fill the entire height of the screen.
+			 *
+			 * Normally, the popup will flex in height to match the size of the content until the
+			 * screen bounds are met. Use this if you require a full-height popup but don't have
+			 * enough content to fill the space or you want a component to stretch to the edges.
+			 *
+			 * @type {Boolean}
+			 * @public
+			 */
+			fullHeight: PropTypes.bool,
 
 			/**
 			 * Function that generates unique identifiers for Panel instances
@@ -156,6 +168,7 @@ const PopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		},
 
 		defaultProps: {
+			fullHeight: false,
 			index: 0,
 			noAnimation: false,
 			position: 'right',
@@ -168,7 +181,7 @@ const PopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		},
 
 		computed: {
-			className: ({width, styler}) => styler.append(width)
+			className: ({fullHeight, width, styler}) => styler.append(width, {fullHeight})
 		},
 
 		render: ({children, className, generateId, id, index, noAnimation, onBack, onClose, ...rest}) => {
@@ -187,11 +200,12 @@ const PopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				}
 			}
 
+			delete rest.fullHeight;
 			delete rest.width;
 
 			return (
 				<Popup {...popupProps} className={className} data-index={index} id={id} css={css} noAnimation={noAnimation} onClose={onClose}>
-					<Panels
+					<CancelableWrapped
 						{...rest}
 						arranger={panelArranger}
 						className={css.viewport}
@@ -204,7 +218,7 @@ const PopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 						type={panelType}
 					>
 						{children}
-					</Panels>
+					</CancelableWrapped>
 				</Popup>
 			);
 		}